pyproject.toml
[project]
name = "cloude-storage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.135.2",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
]
